walidacja czesc pierwsza

+++ src/main/java/com/bkpw/projektkoncowy/service/UserService.java
@@ -1,12 +1,18 @@
 package com.bkpw.projektkoncowy.service;
 
 import com.bkpw.projektkoncowy.entity.User;
+import com.bkpw.projektkoncowy.exception.BindingResultException;
 import com.bkpw.projektkoncowy.exception.NotFoundException;
 import com.bkpw.projektkoncowy.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.crypto.factory.PasswordEncoderFactories;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
 
 import java.util.List;
 import java.util.Optional;
@@ -17,10 +23,22 @@ public class UserService implements GenericService<User> {
     @Autowired
     UserRepository userRepository;
 
+
+    public User create(User user, BindingResult bindingResult) {
+        validate(user, null, bindingResult);
+        return create(user);
+    }
+
     @Override
     public User create(User user) {
+        String newPassword = user.getPassword();
+        if (StringUtils.hasText(newPassword)) {
+            PasswordEncoder encoder = PasswordEncoderFactories.createDelegatingPasswordEncoder();
+            user.setPassword(encoder.encode(newPassword));
+        }
         return userRepository.save(user);
     }
+
 
     @Override
     public User getOne(Long id) {
@@ -39,17 +57,34 @@ public class UserService implements GenericService<User> {
 
     @Override
     public void delete(Long id) {
+        if (userRepository.existsById(id)) {
+            throw new NotFoundException(String.format("User with id: %s not found", id));
+        }
         userRepository.deleteById(id);
     }
 
     @Override
-    public User update(User user, Long id) {
+    public User update(User user, Long id, BindingResult bindingResult) {
+        validate(user, getOne(id).getName(), bindingResult);
         Optional<User> OptionalUserToUpdate = userRepository.findById(id);
         if (!OptionalUserToUpdate.isPresent()) {
             throw new NotFoundException(String.format("User with id %s not found", id));
         }
         user.setId(id);
         return userRepository.save(user);
+    }
+
+
+    private void validate(User user, String currentUsername, BindingResult bindingResult) {
+        if (!user.getName().equals(currentUsername)
+                && userRepository.existsByUsername(user.getName())) {
+            bindingResult.addError(
+                    new FieldError("user", "field",
+                            String.format("User with username %s already exists", user.getName())));
+        }
+        if (bindingResult.hasErrors()) {
+            throw new BindingResultException(bindingResult);
+        }
     }
